// server/types/douyu.d.ts
// Generated by https://quicktype.io

export interface DouyuParam {
    sign: string,
    cdn: string,
    rate: number
    ver: string,
    iar: number,
    ive: number,
    hevc: number,
    aid: string,
    uid: string,
    fa: number
}

export interface DouyuData {
    error: number;
    msg: string;
    data: Data;
}

export interface Data {
    room_id: number;
    is_mixed: boolean;
    mixed_live: string;
    mixed_url: string;
    rtmp_cdn: string;
    rtmp_url: string;
    rtmp_live: string;
    client_ip: string;
    inNA: number;
    rateSwitch: number;
    rate: number;
    cdnsWithName: CdnsWithName[];
    multirates: Multirate[];
    isPassPlayer: number;
    eticket: null;
    online: number;
    mixedCDN: string;
    p2p: number;
    streamStatus: number;
    smt: number;
    p2pMeta: P2PMeta;
    p2pCid: number;
    p2pCids: string;
    player_1: string;
    h265_p2p: number;
    h265_p2p_cid: number;
    h265_p2p_cids: string;
    acdn: string;
    av1_url: string;
    rtc_stream_url: string;
    rtc_stream_config: string;
    pictureQualitySwitch: number;
    show_id: string;
}

export interface CdnsWithName {
    name: string;
    cdn: string;
    isH265: boolean;
}

export interface Multirate {
    name: string;
    rate: number;
    highBit: number;
    bit: number;
    diamondFan: number;
}

export interface P2PMeta {
    best_play_buffer_ms: number;
    dns_tmoms: number;
    fake_peer_factor: number;
    httpdns_aging_sec: number;
    httpdns_failnum_max: number;
    max_play_buffer_ms: number;
    min_play_buffer_ms: number;
    min_play_delay_time_second: number;
    pcdn: string;
    pcdn_table: string;
    play_fast_scale: number;
    play_slow_scale: number;
    sdkapi_iplist: string;
    sdkapi_ipv6list: string;
    stream_props: StreamProp[];
    use_p2p_agent: number;
    using_ali_httpdns: number;
    using_httpdns: number;
    using_syshost: number;
    using_tct_httpdns: number;
    xp2p_api_domain: string;
    xp2p_domain: string;
    xp2p_emergency_wnd_ms: number;
    xp2p_enable_check_source_id: number;
    xp2p_enable_repull_full_stream: number;
    xp2p_enable_repull_full_stream_when_source_changed2: number;
    xp2p_enable_websocket: number;
    xp2p_max_cache_slice_num: number;
    xp2p_subs_domain: string;
    xp2p_txDelay: number;
    xp2p_txExpire: number;
    xp2p_txSecret: string;
    xp2p_txTime: string;
    zero_package_domain: string;
}

export interface StreamProp {
    sid: string;
    txSecret: string;
}